+++ database/migrations/2018_07_22_002000_create_love_reactions_table.php
@@ -11,6 +11,10 @@
 
 declare(strict_types=1);
 
+use Cog\Laravel\Love\Reactant\Models\Reactant;
+use Cog\Laravel\Love\Reacter\Models\Reacter;
+use Cog\Laravel\Love\Reaction\Models\Reaction;
+use Cog\Laravel\Love\ReactionType\Models\ReactionType;
 use Cog\Laravel\Love\Support\Database\Migration;
 use Illuminate\Database\Schema\Blueprint;
 
@@ -23,7 +27,7 @@ final class CreateLoveReactionsTable extends Migration
      */
     public function up(): void
     {
-        $this->schema->create('love_reactions', function (Blueprint $table) {
+        $this->schema->create((new Reaction())->getTable(), function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('reactant_id');
             $table->unsignedBigInteger('reacter_id');
@@ -52,17 +56,17 @@ final class CreateLoveReactionsTable extends Migration
             $table
                 ->foreign('reactant_id')
                 ->references('id')
-                ->on('love_reactants')
+                ->on((new Reactant())->getTable())
                 ->onDelete('cascade');
             $table
                 ->foreign('reacter_id')
                 ->references('id')
-                ->on('love_reacters')
+                ->on((new Reacter())->getTable())
                 ->onDelete('cascade');
             $table
                 ->foreign('reaction_type_id')
                 ->references('id')
-                ->on('love_reaction_types')
+                ->on((new ReactionType())->getTable())
                 ->onDelete('cascade');
         });
     }
@@ -74,6 +78,6 @@ final class CreateLoveReactionsTable extends Migration
      */
     public function down(): void
     {
-        $this->schema->dropIfExists('love_reactions');
+        $this->schema->dropIfExists((new Reaction())->getTable());
     }
 }