+++ database/migrations/2018_07_25_001000_create_love_reactant_reaction_totals_table.php
@@ -11,6 +11,7 @@
 
 declare(strict_types=1);
 
+use Cog\Laravel\Love\Reactant\ReactionTotal\Models\ReactionTotal;
 use Cog\Laravel\Love\Support\Database\Migration;
 use Illuminate\Database\Schema\Blueprint;
 
@@ -23,7 +24,7 @@ final class CreateLoveReactantReactionTotalsTable extends Migration
      */
     public function up(): void
     {
-        $this->schema->create('love_reactant_reaction_totals', function (Blueprint $table) {
+        $this->schema->create((new ReactionTotal())->getTable(), function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('reactant_id');
             $table->unsignedBigInteger('count');
@@ -45,6 +46,6 @@ final class CreateLoveReactantReactionTotalsTable extends Migration
      */
     public function down(): void
     {
-        $this->schema->dropIfExists('love_reactant_reaction_totals');
+        $this->schema->dropIfExists((new ReactionTotal())->getTable());
     }
 }