+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to `laravel-love` will be documented in this file.
 
 ## [Unreleased]
 
+## [8.4.0]
+
 ### Added
 
 - ([#165]) Added table names configuration
@@ -436,7 +438,8 @@ Follow [upgrade instructions](UPGRADING.md#from-v5-to-v6) to migrate database to
 
 - Initial release
 
-[Unreleased]: https://github.com/cybercog/laravel-love/compare/8.3.1...master
+[Unreleased]: https://github.com/cybercog/laravel-love/compare/8.4.0...master
+[8.4.0]: https://github.com/cybercog/laravel-love/compare/8.3.1...8.4.0
 [8.3.1]: https://github.com/cybercog/laravel-love/compare/8.3.0...8.3.1
 [8.3.0]: https://github.com/cybercog/laravel-love/compare/8.2.0...8.3.0
 [8.2.0]: https://github.com/cybercog/laravel-love/compare/8.1.2...8.2.0